Fix minor error in policy path
The section _Saving Configurations to Source Control_ outlines the steps necessary to download template and policy configuration as `.content.xml` files. The path to policies is different from the one in the project structure. I believe it should be `/conf/wknd/settings/wcm/policies` (`policies` is a sibling of `templates`, not its child).