ExL syntax cleanup (#393)
* ExL syntax cleanup

Please check this PR before merging. Most of the changes I made are benign, like removing spaces to get notes to work. But I also removed {.line-numbers} from code blocks. We don't support that.

Reject the PR if you want to track the {.line-numbers} in comments or some other way.

* Update help/forms/storing-retrieving-form-data/part4.md