Update integrate-spa.md
When trying out the "Bonus - Mock JSON API" steps, after changing the .env.development variables to point to local mock JSON content instead of AEM's en.model.json, and after restarting the webpack server, reloading shows only an empty space and a "http://localhost:4502/mock-content/mock.model.json not found" message in the console.  Commenting out the path to the API Host variable is necessary in order to see the live changes in the mock.model.json reflected in the page.