Update set-up-a-local-aem-development-environment.md
Provided an edit, relating logic and language. This edit was not solicited. I was reading the webpage and found myself a little distracted by the content, and wondering if I would be allowed to edit this page. I am interested to know the Adobe Style Guide, so may send that to me at: qdp@qmberli.com. I have great respect for Adobe and I want to see more people using their products.

I would be glad to help edit this content and expand the review, based on criteria and language used or style. However, I will stop here. 
Best regards,
K. Dianne